+++ gplugins/__init__.py
@@ -1,6 +1,6 @@
 """gplugins - gdsfactory plugins."""
 
-__version__ = "1.3.8"
+__version__ = "1.4.0"
 
 import pathlib